Anthropic: rename wiretypes
